Use correct ports (B device port is receive)
