+++ src/TipsyRecvText.cpp
@@ -2,8 +2,8 @@
 // Created by Paul Walker on 7/14/23.
 //
 
-#include "TipsyProto.hpp"
 #include "TipsyWidgetBits.h"
+#include "components.hpp"
 #include <tipsy/tipsy.h>
 
 struct TipsyRecvText : rack::Module
@@ -90,12 +90,6 @@ struct TxtOut : rack::LedDisplayTextField
     }
 };
 
-struct USB_A_Port : rack::app::SvgPort {
-	USB_A_Port() {
-		rack::app::SvgPort::setSvg(rack::window::Svg::load(rack::asset::plugin(pluginInstance, "res/USB-A.svg")));
-	}
-};
-
 struct TipsyRecvTextWidget : rack::ModuleWidget
 {
     TipsyRecvTextWidget(TipsyRecvText *m) {
@@ -114,7 +108,7 @@ struct TipsyRecvTextWidget : rack::ModuleWidget
         addChild(ti);
 
         addInput(
-            rack::createInput<USB_A_Port>(rack::Vec(40, RACK_HEIGHT - 40), module, TipsyRecvText::TXT_IN));
+            rack::createInput<USB_B_Port>(rack::Vec(box.size.x - 50, RACK_HEIGHT - 40), module, TipsyRecvText::TXT_IN));
 
 
     }
